requirements.txt
alembic
anndata
black >= 20.8b1 , < 21.0
botocore>=1.14.17
boto3>=1.11.17
chalice==1.12.0
coverage>=2.0.15
flake8-black
furl
jsonschema
moto==1.3.14
numpy
owlready2
pandas
PyMySQL==0.9.3
pytest
requests>=2.22.0
s3fs==0.4.2
scanpy
pyyaml
-r ./backend/chalice/api_server/requirements.txt
